requirements.txt
torch
transformers
sentence-transformers
faiss-cpu
numpy
rank-bm25
streamlit
